+++ app/Http/Controllers/Manager/AuthController.php
@@ -25,6 +25,7 @@ use Adshares\Adserver\Mail\AuthRecovery;
 use Adshares\Adserver\Mail\UserEmailActivate;
 use Adshares\Adserver\Mail\UserEmailChangeConfirm1Old;
 use Adshares\Adserver\Mail\UserEmailChangeConfirm2New;
+use Adshares\Adserver\Models\Config;
 use Adshares\Adserver\Models\Token;
 use Adshares\Adserver\Models\User;
 use Illuminate\Http\JsonResponse;
@@ -74,16 +75,27 @@ class AuthController extends Controller
         Validator::make($request->all(), ['user.email_confirm_token' => 'required'])->validate();
 
         DB::beginTransaction();
-        if (false === $token = Token::check($request->input('user.email_confirm_token'))) {
+        $token = Token::check($request->input('user.email_confirm_token'));
+        if (false === $token) {
             return self::json([], Response::HTTP_FORBIDDEN);
         }
+
+        /** @var User $user */
         $user = User::find($token['user_id']);
+
         if (empty($user)) {
             return self::json([], Response::HTTP_FORBIDDEN);
         }
-        $user->email_confirmed_at = date('Y-m-d H:i:s');
+
+        $user->confirmEmail();
+
         $user->save();
+
         DB::commit();
+
+        if (Config::isTrueOnly(Config::BONUS_NEW_USER_ENABLED)) {
+            $user->awardBonus(Config::fetchInt(Config::BONUS_NEW_USER_AMOUNT));
+        }
 
         return self::json($user->toArray());
     }
@@ -215,7 +227,7 @@ class AuthController extends Controller
     /**
      * Log the user out of the application.
      *
-     * @param \Illuminate\Http\Request $request
+     * @param Request $request
      *
      * @return JsonResponse
      */
@@ -248,7 +260,7 @@ class AuthController extends Controller
     /**
      * Start password recovery process - generate and send email.
      *
-     * @param \Illuminate\Http\Request $request
+     * @param Request $request
      *
      * @return JsonResponse
      */