+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Bonus credits for new users (on email confirmation)
 
 ## [0.5.4] - 2019-04-12
 ### Fixed
@@ -72,7 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v0.5.4...master
+[Unreleased]: https://github.com/adshares/adserver/compare/v0.5.4...develop
 [0.5.4]: https://github.com/adshares/adserver/compare/v0.5.3...v0.5.4
 [0.5.3]: https://github.com/adshares/adserver/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/adshares/adserver/compare/v0.5.1...v0.5.2